Adding a temp header, fixing tailwind (#9)
Adding a header mostly for checking vue components and app.
Fixing tailwind, did not work on my local and I assume it did not work elsewhere.
Array type in header in Sanity was missing required "of", arrays must be specified what the arrays are of. I assumed it was supposed to to show companies.